requirements.txt
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
babel==2.16.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
crispy-bootstrap4==2023.1
cryptography==45.0.5
cycler==0.12.1
DateTime==5.5
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.2
django-allauth==64.0.0
django-crispy-forms==2.1
django-dotenv==1.4.2
django-forms-bootstrap==3.1.0
django-mathfilters==1.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-spectacular==0.27.2
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.0.1
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.14.0
fonttools==4.53.0
fqdn==1.5.1
greenlet==3.2.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.7
inflection==0.5.1
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.5.1
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
kmodes==0.12.2
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.0.2
mlxtend==0.23.4
mysql-connector-python==8.4.0
mysqlclient==2.2.4
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.1
notebook_shim==0.2.4
numpy==1.26.4
openpyxl==3.1.5
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pillow==10.2.0
platformdirs==4.2.2
prometheus_client==0.21.1
prompt_toolkit==3.0.46
psutil==5.9.8
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2024.1
pywin32==306
pywinpty==2.0.14
PyYAML==6.0.1
pyzmq==26.0.3
qrcode==7.4.2
referencing==0.35.1
req==1.0.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==70.0.0
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.47.1
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.35.0
virtualenv==20.26.2
VirtualEnvironment==0.1.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
XlsxWriter==3.2.3
zope.interface==6.4.post2